fix DBL_MAX

+++ ft_get_arg_f.c
@@ -27,7 +27,7 @@ int ft_get_arg_f(t_spec *s_args, va_list args)
 		nb = (double)va_arg(args, double);
 //	if (nb == "NaN")
 	ft_putsign(s_args);
-	ret = ft_put_f_str((float)nb, s_args);
+	ret = ft_put_f_str((long double)nb, s_args);
 	ft_get_len_output_f(s_args);
 	ret = (!ret ? ft_putoutput_f(s_args) : ret);
